Use Decimal, better for money and use main.py

+++ payment_methods/google_pay_payment.py
@@ -1,5 +1,7 @@
+from decimal import Decimal
 from .payment import Payment
 
+
 class GooglePayPayment(Payment):
-    def pay(self, amount: float):
+    def pay(self, amount: Decimal):
         print(f"Successfully paid ${amount} to merchant using Google Pay")
