+++ payment_methods/payment.py
@@ -1,7 +1,9 @@
 from abc import ABC, abstractmethod
+from decimal import Decimal
+
 
 class Payment(ABC):
 
     @abstractmethod
-    def pay(self, amount: float):
+    def pay(self, amount: Decimal):
         pass
