+++ payment_methods/paypal_payment.py
@@ -1,5 +1,7 @@
+from decimal import Decimal
 from .payment import Payment
 
+
 class PayPalPayment(Payment):
-    def pay(self, amount: float):
+    def pay(self, amount: Decimal):
         print(f"Successfully paid ${amount} to merchant using PayPal")
